Splitted Post.jsx to be more readable
