+++ src/components/AuthorBox.jsx
@@ -6,18 +6,20 @@ import styled from '@emotion/styled'
 import { css } from '@emotion/core'
 
 import getAuthorUrl from '../utils/getAuthorUrl'
-import {palette} from '../utils/colors'
-import ResponsiveFlexBox from '../components/ResponsiveFlexBox'
+import { palette } from '../utils/colors'
+import {
+  SMALL_SCREEN_ONLY,
+  MEDIUM_SCREEN_UP
+} from '../utils/breakpoints'
+import ResponsiveFlexBox from './ResponsiveFlexBox'
 import {
   Instagram,
   Facebook,
   Website,
   Youtube
 } from '../components/SocialLinks'
-import {
-  SMALL_SCREEN_ONLY,
-  MEDIUM_SCREEN_UP
-} from '../utils/breakpoints'
+import DangerousHTML from '../components/DangerousHTML'
+import Card from './Card'
 
 const ImgWrapper = styled.div`
   ${SMALL_SCREEN_ONLY} {
@@ -45,7 +47,7 @@ const socialIcon = css`
 `;
 
 
-function AuthorBox({ author, showProfileLink }) {
+function AuthorBox({ author, showProfileLink, ...otherProps }) {
   const {
     name,
     slug,
@@ -57,81 +59,81 @@ function AuthorBox({ author, showProfileLink }) {
   } = author.frontmatter;
 
   return (
-    <ResponsiveFlexBox className="content">
-      { !avatar ? null :
-        <ImgWrapper>
-          <Img
-            fluid={avatar.childImageSharp.fluid}
-            alt={avatar.childImageSharp.alt_text}
-          />
-        </ImgWrapper>
-      }
-      <div>
-        <h1>{ name }</h1>
-        <div dangerouslySetInnerHTML={{ __html: author.html }} />
-        { !website ? null :
-          <LinkWrapper>
-            <a
-              href={website}
-              target="_blank"
-              rel="noopener noreferrer"
-              title="Sito web"
-            >
-              <Website css={socialIcon} />
-              { website }
-            </a>
-          </LinkWrapper>
+    <Card itemScope="itemscope" itemType="http://schema.org/Person" {...otherProps}>
+      <ResponsiveFlexBox className="content">
+        { !avatar ? null :
+          <ImgWrapper>
+            <Img
+              fluid={avatar.childImageSharp.fluid}
+              alt={avatar.childImageSharp.alt_text}
+            />
+          </ImgWrapper>
         }
-        { !facebook || !facebook.url ? null :
-          <LinkWrapper>
-            <a
-              href={facebook.url}
-              target="_blank"
-              rel="noopener noreferrer"
-              title="Facebook"
-            >
-              <Facebook css={socialIcon} />
-              { facebook.name }
-            </a>
-          </LinkWrapper>
-        }
-        { !instagram || !instagram.url ? null :
-          <LinkWrapper>
-            <a
-              href={instagram.url}
-              target="_blank"
-              rel="noopener noreferrer"
-              title="Instagram"
-            >
-              <Instagram css={socialIcon} />
-              { instagram.name }
-            </a>
-          </LinkWrapper>
-        }
-        { !youtube || !youtube.url ? null :
-          <LinkWrapper>
-            <a
-              href={youtube.url}
-              target="_blank"
-              rel="noopener noreferrer"
-              title="YouTube"
-            >
-              <Youtube css={socialIcon} />
-              { youtube.name }
-            </a>
-          </LinkWrapper>
-        }
-        { !showProfileLink || !slug ? null :
-          <p>
-            <Link
-              to={getAuthorUrl(slug)}
-            >
-              Tutti i post di {name}
-            </Link>
-          </p>
-        }
-      </div>
-    </ResponsiveFlexBox>
+        <div>
+          <h1>{ name }</h1>
+          <DangerousHTML html={ author.html } />
+          { !website ? null :
+            <LinkWrapper>
+              <a
+                href={website}
+                target="_blank"
+                rel="noopener noreferrer"
+                title="Sito web"
+              >
+                <Website css={socialIcon} />
+                { website }
+              </a>
+            </LinkWrapper>
+          }
+          { !facebook || !facebook.url ? null :
+            <LinkWrapper>
+              <a
+                href={facebook.url}
+                target="_blank"
+                rel="noopener noreferrer"
+                title="Facebook"
+              >
+                <Facebook css={socialIcon} />
+                { facebook.name }
+              </a>
+            </LinkWrapper>
+          }
+          { !instagram || !instagram.url ? null :
+            <LinkWrapper>
+              <a
+                href={instagram.url}
+                target="_blank"
+                rel="noopener noreferrer"
+                title="Instagram"
+              >
+                <Instagram css={socialIcon} />
+                { instagram.name }
+              </a>
+            </LinkWrapper>
+          }
+          { !youtube || !youtube.url ? null :
+            <LinkWrapper>
+              <a
+                href={youtube.url}
+                target="_blank"
+                rel="noopener noreferrer"
+                title="YouTube"
+              >
+                <Youtube css={socialIcon} />
+                { youtube.name }
+              </a>
+            </LinkWrapper>
+          }
+          { !showProfileLink || !slug ? null :
+            <p>
+              <Link to={ getAuthorUrl(slug) }>
+                Tutti i post di {name}
+              </Link>
+            </p>
+          }
+        </div>
+      </ResponsiveFlexBox>
+    </Card>
   )
 }
