+++ src/components/post/index.js
@@ -1,0 +1,7 @@
+export { default as Article } from './Article'
+export { default as PostMeta } from './PostMeta'
+export { default as Vote } from './Vote'
+export { default as TagList } from './TagList'
+export { default as NextPrev } from './NextPrev'
+export { default as FeaturedMedia } from './FeaturedMedia'
+export { default as AttributesTable } from './AttributesTable'
